pyproject.toml
[project]
name = "fabric-utilities"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "azure-core>=1.36.0",
    "azure-identity>=1.25.1",
    "deltalake==0.18.2",
    "polars==1.6.0",
]
